+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -1,4 +1,4 @@
-# pylint: disable=invalid-name
+# pylint: disable=invalid-name,access-member-before-definition
 # pylint: disable=C0302,W0212,W0201
 """
 This module implements the dish manager device for DishLMC.
@@ -21,6 +21,7 @@ from tango import AttrWriteType, DebugIt, DevFloat, DevString, DispLevel
 from tango.server import attribute, command, device_property, run
 
 from ska_mid_dish_manager.component_managers.dish_manager_cm import DishManagerComponentManager
+from ska_mid_dish_manager.component_managers.tango_device_cm import LostConnection
 from ska_mid_dish_manager.interface.input_validation import (
     TrackLoadTableFormatting,
     TrackTableTimestampError,
@@ -73,6 +74,7 @@ class DishManager(SKAController):
             self.logger,
             self._command_tracker,
             self._update_connection_state_attrs,
+            self._attr_quality_state_changed,
             self.get_name(),
             self.DSDeviceFqdn,
             self.SPFDeviceFqdn,
@@ -157,6 +159,19 @@ class DishManager(SKAController):
                 "dsConnectionState",
                 self.component_manager.sub_component_managers["DS"].communication_state,
             )
+
+    def _attr_quality_state_changed(self, attribute_name, new_attribute_quality):
+        # Do not modify or push quality changes before initialization complete
+        if not hasattr(self, "_component_state_attr_map"):
+            self.logger.warning("Init not completed, rejecting attribute quality update")
+            return
+
+        device_attribute_name = self._component_state_attr_map.get(attribute_name, None)
+        if device_attribute_name:
+            attribute_object = getattr(self, device_attribute_name, None)
+            if attribute_object:
+                if attribute_object.get_quality() is not new_attribute_quality:
+                    attribute_object.set_quality(new_attribute_quality, True)
 
     # pylint: disable=unused-argument
     def _component_state_changed(self, *args, **kwargs):
@@ -263,6 +278,7 @@ class DishManager(SKAController):
                 "pointingstate": "pointingState",
                 "configuredband": "configuredBand",
                 "achievedtargetlock": "achievedTargetLock",
+                "configuretargetlock": "configureTargetLock",
                 "healthstate": "healthState",
                 "b1capabilitystate": "b1CapabilityState",
                 "b2capabilitystate": "b2CapabilityState",
@@ -275,7 +291,10 @@ class DishManager(SKAController):
                 "achievedpointing": "achievedPointing",
                 "achievedpointingaz": "achievedPointingAz",
                 "achievedpointingel": "achievedPointingEl",
+                "band1pointingmodelparams": "band1PointingModelParams",
                 "band2pointingmodelparams": "band2PointingModelParams",
+                "band3pointingmodelparams": "band3PointingModelParams",
+                "band4pointingmodelparams": "band4PointingModelParams",
                 "attenuationpolh": "attenuationPolH",
                 "attenuationpolv": "attenuationPolV",
                 "kvalue": "kValue",
@@ -324,7 +343,6 @@ class DishManager(SKAController):
                 "band5aSamplerFrequency",
                 "band5bSamplerFrequency",
                 "capturing",
-                "configureTargetLock",
                 "dshMaxShortTermPower",
                 "dshPowerCurtailment",
                 "frequencyResponse",
@@ -414,6 +432,7 @@ class DishManager(SKAController):
         doc="Indicates whether the Dish is on target or not based on the "
         "pointing error and time period parameters defined in "
         "configureTargetLock.",
+        access=AttrWriteType.READ,
     )
     def achievedTargetLock(self):
         """Returns the achievedTargetLock"""
@@ -474,10 +493,21 @@ class DishManager(SKAController):
 
     @attribute(
         dtype=(DevFloat,),
-        max_dim_x=5,
+        max_dim_x=20,
         access=AttrWriteType.READ_WRITE,
-        doc="Parameters for (local) Band 1 pointing models used by Dish to do "
-        "pointing corrections.",
+        doc="""
+            Parameters for (local) Band 1 pointing models used by Dish to do pointing corrections.
+
+            When writing to this attribute, the selected band for correction will be set to B1.
+
+            Band 1 pointing model parameters are:
+            [0] IA, [1] CA, [2] NPAE, [3] AN, [4] AN0, [5] AW, [6] AW0, [7] ACEC, [8] ACES,
+            [9] ABA, [10] ABphi, [11] CAobs, [12] IE, [13] ECEC, [14] ECES, [15] HECE4,
+            [16] HESE4, [17] HECE8, [18] HESE8, [19] Eobs
+
+            When writing we expect a list of 2 values. Namely, CAobs and Eobs. Only those two
+            values will be updated.
+        """,
     )
     def band1PointingModelParams(self):
         """Returns the band1PointingModelParams"""
@@ -486,10 +516,21 @@ class DishManager(SKAController):
     @band1PointingModelParams.write
     def band1PointingModelParams(self, value):
         """Set the band1PointingModelParams"""
-        # pylint: disable=attribute-defined-outside-init
-        self._band1_pointing_model_params = value
-        self.push_change_event("band1PointingModelParams", value)
-        self.push_archive_event("band1PointingModelParams", value)
+        self.logger.debug("band1PointingModelParams write method called with params %s", value)
+
+        # The argument value is a list of two floats: [off_xel, off_el]
+        if len(value) != 2:
+            raise ValueError(f"Length of argument ({len(value)}) is not as expected (2).")
+
+        if hasattr(self, "component_manager"):
+            if "DS" in self.component_manager.sub_component_managers:
+                try:
+                    ds_com_man = self.component_manager.sub_component_managers["DS"]
+                    ds_com_man.write_attribute_value("band1PointingModelParams", value)
+                except tango.DevFailed:
+                    self.logger.exception("Could not reach DS to write band1PointingModelParams")
+        else:
+            self.logger.warning("No component manager to write band1PointingModelParams yet")
 
     @attribute(
         dtype=(DevFloat,),
@@ -497,6 +538,8 @@ class DishManager(SKAController):
         access=AttrWriteType.READ_WRITE,
         doc="""
             Parameters for (local) Band 2 pointing models used by Dish to do pointing corrections.
+
+            When writing to this attribute, the selected band for correction will be set to B2.
 
             Band 2 pointing model parameters are:
             [0] IA, [1] CA, [2] NPAE, [3] AN, [4] AN0, [5] AW, [6] AW0, [7] ACEC, [8] ACES,
@@ -532,10 +575,21 @@ class DishManager(SKAController):
 
     @attribute(
         dtype=(DevFloat,),
-        max_dim_x=5,
+        max_dim_x=20,
         access=AttrWriteType.READ_WRITE,
-        doc="Parameters for (local) Band 3 pointing models used by Dish to do "
-        "pointing corrections.",
+        doc="""
+            Parameters for (local) Band 3 pointing models used by Dish to do pointing corrections.
+
+            When writing to this attribute, the selected band for correction will be set to B3.
+
+            Band 3 pointing model parameters are:
+            [0] IA, [1] CA, [2] NPAE, [3] AN, [4] AN0, [5] AW, [6] AW0, [7] ACEC, [8] ACES,
+            [9] ABA, [10] ABphi, [11] CAobs, [12] IE, [13] ECEC, [14] ECES, [15] HECE4,
+            [16] HESE4, [17] HECE8, [18] HESE8, [19] Eobs
+
+            When writing we expect a list of 2 values. Namely, CAobs and Eobs. Only those two
+            values will be updated.
+        """,
     )
     def band3PointingModelParams(self):
         """Returns the band3PointingModelParams"""
@@ -544,17 +598,39 @@ class DishManager(SKAController):
     @band3PointingModelParams.write
     def band3PointingModelParams(self, value):
         """Set the band3PointingModelParams"""
-        # pylint: disable=attribute-defined-outside-init
-        self._band3_pointing_model_params = value
-        self.push_change_event("band3PointingModelParams", value)
-        self.push_archive_event("band3PointingModelParams", value)
+        self.logger.debug("band3PointingModelParams write method called with params %s", value)
+
+        # The argument value is a list of two floats: [off_xel, off_el]
+        if len(value) != 2:
+            raise ValueError(f"Length of argument ({len(value)}) is not as expected (2).")
+
+        if hasattr(self, "component_manager"):
+            if "DS" in self.component_manager.sub_component_managers:
+                try:
+                    ds_com_man = self.component_manager.sub_component_managers["DS"]
+                    ds_com_man.write_attribute_value("band3PointingModelParams", value)
+                except tango.DevFailed:
+                    self.logger.exception("Could not reach DS to write band3PointingModelParams")
+        else:
+            self.logger.warning("No component manager to write band3PointingModelParams yet")
 
     @attribute(
         dtype=(DevFloat,),
-        max_dim_x=5,
+        max_dim_x=20,
         access=AttrWriteType.READ_WRITE,
-        doc="Parameters for (local) Band 4 pointing models used by Dish to do "
-        "pointing corrections.",
+        doc="""
+            Parameters for (local) Band 4 pointing models used by Dish to do pointing corrections.
+
+            When writing to this attribute, the selected band for correction will be set to B4.
+
+            Band 4 pointing model parameters are:
+            [0] IA, [1] CA, [2] NPAE, [3] AN, [4] AN0, [5] AW, [6] AW0, [7] ACEC, [8] ACES,
+            [9] ABA, [10] ABphi, [11] CAobs, [12] IE, [13] ECEC, [14] ECES, [15] HECE4,
+            [16] HESE4, [17] HECE8, [18] HESE8, [19] Eobs
+
+            When writing we expect a list of 2 values. Namely, CAobs and Eobs. Only those two
+            values will be updated.
+        """,
     )
     def band4PointingModelParams(self):
         """Returns the band4PointingModelParams"""
@@ -563,10 +639,21 @@ class DishManager(SKAController):
     @band4PointingModelParams.write
     def band4PointingModelParams(self, value):
         """Set the band4PointingModelParams"""
-        # pylint: disable=attribute-defined-outside-init
-        self._band4_pointing_model_params = value
-        self.push_change_event("band4PointingModelParams", value)
-        self.push_archive_event("band4PointingModelParams", value)
+        self.logger.debug("band4PointingModelParams write method called with params %s", value)
+
+        # The argument value is a list of two floats: [off_xel, off_el]
+        if len(value) != 2:
+            raise ValueError(f"Length of argument ({len(value)}) is not as expected (2).")
+
+        if hasattr(self, "component_manager"):
+            if "DS" in self.component_manager.sub_component_managers:
+                try:
+                    ds_com_man = self.component_manager.sub_component_managers["DS"]
+                    ds_com_man.write_attribute_value("band4PointingModelParams", value)
+                except tango.DevFailed:
+                    self.logger.exception("Could not reach DS to write band4PointingModelParams")
+        else:
+            self.logger.warning("No component manager to write band4PointingModelParams yet")
 
     @attribute(
         dtype=(DevFloat,),
@@ -739,8 +826,8 @@ class DishManager(SKAController):
         """Set the configureTargetLock"""
         # pylint: disable=attribute-defined-outside-init
         self._configure_target_lock = value
-        self.push_change_event("configureTargetLock", value)
-        self.push_archive_event("configureTargetLock", value)
+        ds_com_man = self.component_manager.sub_component_managers["DS"]
+        ds_com_man.write_attribute_value("configureTargetLock", value)
 
     @attribute(
         max_dim_x=2,
@@ -1099,8 +1186,12 @@ class DishManager(SKAController):
                     try:
                         spfrx_com_man = self.component_manager.sub_component_managers["SPFRX"]
                         spfrx_com_man.execute_command("MonitorPing", None)
+                    except LostConnection:
+                        self.logger.error(
+                            "Could not connect to [%s] for MonitorPing", self.SPFRxDeviceFqdn
+                        )
                     except tango.DevFailed:
-                        self.logger.debug("Could not reach SPFRx")
+                        pass
 
     # pylint: disable=too-few-public-methods
     class AbortCommandsCommand(SlowCommand):
@@ -1431,22 +1522,21 @@ class DishManager(SKAController):
 
     @command(
         dtype_in="DevVarFloatArray",
-        doc_in="[0]: Azimuth\n[1]: Elevation,\n[2]: Azimuth Speed,\n[3]: Elevation Speed",
+        doc_in="[0]: Azimuth\n[1]: Elevation",
         dtype_out="DevVarLongStringArray",
         display_level=DispLevel.OPERATOR,
     )
     def Slew(self, values):  # pylint: disable=unused-argument
         """
-        Trigger the Dish to start moving at the given speeds to the commanded (Az,El) position.
+        Trigger the Dish to start moving to the commanded (Az,El) position.
 
-        :param argin: the az, el, az speed, and el speed for the pointing in stringified json
-            format
+        :param argin: the az, el for the pointing in stringified json format
 
         :return: A tuple containing a return code and a string
             message indicating status.
         """
-        if len(values) != 4:
-            raise ValueError(f"Length of argument ({len(values)}) is not as expected (4).")
+        if len(values) != 2:
+            raise ValueError(f"Length of argument ({len(values)}) is not as expected (2).")
 
         handler = self.get_command_object("Slew")
         result_code, unique_id = handler(values)
@@ -1517,6 +1607,20 @@ class DishManager(SKAController):
     @command(  # type: ignore[misc]
         dtype_in="DevVarFloatArray",
         dtype_out="DevVarLongStringArray",
+        doc_in="""
+            Load the static offsets for the currently selected band for correction.
+
+            Pointing model parameters are:
+            [0] IA, [1] CA, [2] NPAE, [3] AN, [4] AN0, [5] AW, [6] AW0, [7] ACEC, [8] ACES,
+            [9] ABA, [10] ABphi, [11] CAobs, [12] IE, [13] ECEC, [14] ECES, [15] HECE4,
+            [16] HESE4, [17] HECE8, [18] HESE8, [19] Eobs
+
+            Note: To change the currently selected band for correction to B<N>, write to the
+            band<N>PointingModelParams attribute.
+
+            When writing we expect a list of 2 values. Namely, CAobs and Eobs. Only those two
+            values will be updated.
+        """,
     )
     @DebugIt()  # type: ignore[misc]
     def TrackLoadStaticOff(self, values) -> DevVarLongStringArrayType:
@@ -1572,7 +1676,9 @@ class DishManager(SKAController):
     )
     def SetKValue(self, value) -> DevVarLongStringArrayType:
         """
-        This command sets the kValue on SPFRx
+        This command sets the kValue on SPFRx.
+        Note that it will only take effect after
+        SPFRx has been restarted.
         """
         handler = self.get_command_object("SetKValue")
         return_code, message = handler(value)