+++ tests/acceptance/test_capability_state.py
@@ -1,4 +1,5 @@
 """Test CapabilityState"""
+
 import pytest
 import tango