+++ tests/acceptance/test_ds_component_manager.py
@@ -1,4 +1,5 @@
 """Test DS component manager"""
+
 import logging
 from threading import Lock